solve previous databse error
change coloums name to mathc mysql's rule.
ex: "lot title" ---> "lot_title"
