+++ webdev-challenge/app/Table.php
@@ -3,9 +3,20 @@
 namespace App;
 
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\DB;
 
 class Table extends Model
 {
 
+    public function fetchData()
+    {
 
+        $content = DB::select('SELECT  date_format(date,\'%Y-%m\') as yyyymm, category, SUM(tax_amount) as amount
+                                            FROM Data
+                                            WHERE date_format(date,\'%Y-%m\') < date_format(now(),\'%Y-%m\') 
+                                            GROUP BY category, date_format(date,\'%Y-%m\')
+                                            ORDER BY category, yyyymm DESC;');
+
+        return [table_content => $content];
+    }
 }
